move helpers, undo htf change
