+++ openfe/protocols/openmm_rfe/_rfe_utils/relative.py
@@ -1772,7 +1772,7 @@ class HybridTopologyFactory:
                 # with zeroed out parameters; add old charge to
                 # standard_nonbonded and zero sterics
                 check_index = self._hybrid_system_forces['standard_nonbonded_force'].addParticle(
-                    charge_old, sigma_old, 0.0)
+                    charge_old, 0.5*(sigma_old+sigma_new), 0.0)
                 _check_indices(particle_index, check_index)
 
                 # Charge is charge_old at lambda_electrostatics = 0,
