+++ openfe/tests/conftest.py
@@ -24,7 +24,7 @@ import openfe
 from openfe.protocols.openmm_rfe import RelativeHybridTopologyProtocol
 from openfe.protocols.openmm_rfe._rfe_utils.relative import HybridTopologyFactory
 from openfe.protocols.openmm_septop.utils import deserialize
-from openfe.tests.helpers import make_htf
+from openfe.tests.protocols.openmm_rfe.helpers import make_htf
 
 
 class SlowTests:
